Fix bug that only occurs on a timout 20060628.003 -> SH
git-svn-id: file:///jac_sw/svnroot/scicom/trunk/omp/java/qt@9987 64693ad5-8cdc-4012-857f-4cd9bf50b8c8

+++ src/main/edu/jach/qt/gui/XmlUtils.java
@@ -453,10 +453,12 @@ public class XmlUtils {
 	return data;
     }
 
-    public static void clearProjectData() {
-	File pdf = new File (System.getProperty("msbSummary")+"."+System.getProperty("user.name"));
-	pdf.delete();
-    }
+    public static void clearProjectData()
+	{
+		File pdf = new File( System.getProperty( "msbSummary" ) + "." + System.getProperty( "user.name" ) );
+		if( pdf.exists() && pdf.canWrite() )
+			pdf.delete();
+	}
 
     public static int getProjectCount() {
 	return getProjectData()[0].size();
